requirements.txt
git+https://github.com/Rapptz/discord.py
pillow
comics
openai
pytz
geopandas
matplotlib
python-dotenv